requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Pydantic
pydantic==2.5.2
pydantic-settings==2.1.0

# RAG and AI
langchain
langchain-core
langchain-community
langchain-text-splitters
langsmith
chromadb==0.4.22
sentence-transformers==2.2.2
openai==1.3.5

# Development
typing-extensions
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
email-validator==2.2.0
